:recycle: #50 ドメイン層にリポジトリのインターフェースを定義
